pyproject.toml
[project]
name = "erl-sc-experiments"
version = "0.1.0"
description = "Equivalence-Aware Risk-Limiting Self-Consistency experiments"
requires-python = ">=3.11"
dependencies = [
    "hydra-core>=1.3.0",
    "omegaconf>=2.3.0",
    "wandb>=0.16.0",
    "datasets>=2.16.0",
    "torch>=2.2.0",
    "transformers>=4.37.0",
    "openai>=1.12.0",
    "anthropic>=0.18.0",
    "numpy>=1.26.0",
    "pandas>=2.2.0",
    "scipy>=1.12.0",
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
    "tqdm>=4.66.0",
    "python-Levenshtein>=0.25.0",
    "rapidfuzz>=3.6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
